NN: Adding utility functions to print, save, load.
The weights and biases of a network can now be saved to a file.
the data is saved in YAML format so it is not dependent on the
language used to code the network.

feedforward can now be called using 'output = net(input)'

import network as n
net = n.Network(struct)
output = net(input)